examples/dvc-hydra/requirements.txt
# python 3.10
gitpython
hydra-core
dvc